Bugfixes in various image transformations.
Fix list:
-Gamma correction
-Monochrome image
-Linear transtormations

+++ src/src/ImgBrightContrast.java
@@ -2,7 +2,7 @@ import java.awt.image.BufferedImage;
 
 public class ImgBrightContrast {
 
-	static public BufferedImage adjustImg(BufferedImage image, Histogram img_hist, int new_brightness, int new_contrast) {
+	static public BufferedImage adjustImg(BufferedImage image, Histogram img_hist, double new_brightness, double new_contrast) {
 		BufferedImage img = new BufferedImage(image.getWidth(), image.getHeight(), image.getType());
 
 		for (int i = 0; i < image.getHeight(); i++) {
@@ -15,9 +15,10 @@ public class ImgBrightContrast {
 				int green = (pixel >> 8) & 0xff;
 				int blue = (pixel) & 0xff;
 				
-				red = (int) (red *(new_contrast / img_hist.dev[0]) + new_brightness - img_hist.med[0] *(new_contrast / img_hist.dev[0]));
-				green = (int) (green *(new_contrast / img_hist.dev[1]) + new_brightness - img_hist.med[1] *(new_contrast / img_hist.dev[1]));
-				blue = (int) (blue *(new_contrast / img_hist.dev[2]) + new_brightness - img_hist.med[2] *(new_contrast / img_hist.dev[2]));
+				
+				red = (int) (red * (new_contrast / img_hist.dev[0]) + new_brightness - img_hist.med[0] * (new_contrast / img_hist.dev[0]));
+				green = (int) (green * (new_contrast / img_hist.dev[1]) + new_brightness - img_hist.med[1] * (new_contrast / img_hist.dev[1]));
+				blue = (int) (blue * (new_contrast / img_hist.dev[2]) + new_brightness - img_hist.med[2] * (new_contrast / img_hist.dev[2]));
 				
 				red = Math.min(red, 255);
 				green = Math.min(green, 255);
